refactor: consolidate test files into __tests__/ directories
